end of perversion with short ints (fixes #1204)
git-svn-id: http://svn.miranda-ng.org/main/trunk@16425 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ protocols/JabberG/src/jabber_proto.h
@@ -573,10 +573,10 @@ struct CJabberProto : public PROTO<CJabberProto>, public IJabberInterface
 
 	void   AuthWorker(MCONTACT hContact, char* authReqType);
 
-	void   UpdatePriorityMenu(short priority);
+	void   UpdatePriorityMenu(int priority);
 
 	HGENMENU m_hMenuPriorityRoot;
-	short  m_priorityMenuVal;
+	int    m_priorityMenuVal;
 	bool   m_priorityMenuValSet;
 
 	//---- jabber_misc.c -----------------------------------------------------------------
